requirements.txt
litellm==1.55.2
python-dotenv==1.0.0
beautifulsoup4==4.12.3
requests==2.31.0
torch==2.5.1
sentence-transformers==3.3.1
hdbscan==0.8.40
tenacity==9.0.0
